wlan-ap-consumer: Added consumer solution specific patches (manual step)
Additional patches required by wlan-ap-consumer. These patches are
applied manually only when building wlan-ap-consumer.

This is a workaround to avoid modifying the trunk of the mainstream
enterprise solution. The workaround could be avoided in future either
by including these patches in the respective feeds on trunk,
or by rearranging the feeds (notably the wifi-trunk feed) differently.

Signed-off-by: Miroslav Rajsek <mrajsek@plume.com>