EXTADT-51: fix optimizer failed by cac timeout
Hostapd deletes and re-adds the AP VIFs on channel switch to
DFS channel.

But due to we don't put bssid in hostapd config, it will let the kernel
mac80111 layer decide the MAC address which would be the first
available MAC address (derive from the radio base address).

Interfaces are removed and re-added only on interface reload
and not the first time interface is added.

Signed-off-by: Jeremy Yang <jyang@plume.com>