EXTADT-80: fix hp_ports by detecting interface change evnet
When the interface is deleted and added back to ovs bridge, the
hp_ports (hairpin ports) are not updated in time. So the traffic
comes and forwads to the same interface would not the hit hairpin
rules. So the inter clients communication would not work.

The patch, by listening interface change events, keeps the deleted
wifi5g port id when the interface is deleted and adds the new wifi5g
port id back to hp_ports when the interface is added.

Signed-off-by: James Wu <jwu@plume.com>