bs: BAND_STEERING_ATTEMPT report is sent to the wrong interface
The BAND_STEERING_ATTEMPT report was sent to the wrong interface (should be 5g instead of 2.4g)
due to !bs_allowed. As a result, osrt could not find BAND_STEERING_ATTEMPT in mqtt messages.

bs_allowed is set by cloud and usually disallowed for 2g and allowed for 5g.

(The patch is from commit 89afa33b)

Signed-off-by: Yi-Chia Hsieh <ychsieh@plume.com>